+++ src/main/java/com/mbel/model/UserEntity.java
@@ -12,7 +12,7 @@ public class UserEntity {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private long userId;
+    private int userId;
     @Column
     private String firstName;
     
@@ -38,12 +38,12 @@ public class UserEntity {
     private Set<RoleEntity> roles;
 
 
-	public long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 
 
-	public void setUserId(long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
